Separate app into authentication and main app functionality
