+++ lib/app/map.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+import 'package:project_mobile_app/app/app_bar.dart';
+
+class MapScreen extends StatefulWidget {
+  const MapScreen({super.key});
+
+  @override
+  State<MapScreen> createState() => _MapScreenState();
+}
+
+class _MapScreenState extends State<MapScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: const PreferredSize(
+        preferredSize: Size.fromHeight(kToolbarHeight),
+        child: MainAppBar(title: 'Map'),
+      ),
+    );
+  }
+}
