+++ lib/authentication/login.dart
@@ -1,17 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:project_mobile_app/login.dart';
+import 'package:project_mobile_app/app/profile.dart';
+import 'package:project_mobile_app/authentication/register.dart';
+import 'package:project_mobile_app/globals.dart' as globals;
 
-const double padding = 10;
-
-class RegisterScreen extends StatefulWidget {
-  const RegisterScreen({super.key});
+class LoginScreen extends StatefulWidget {
+  const LoginScreen({super.key});
 
   @override
-  State<RegisterScreen> createState() => _RegisterScreenState();
+  State<LoginScreen> createState() => _LoginScreenState();
 }
 
-class _RegisterScreenState extends State<RegisterScreen> {
-  String _email = "";
+class _LoginScreenState extends State<LoginScreen> {
   String _username = "";
   String _password = "";
 
@@ -20,35 +19,23 @@ class _RegisterScreenState extends State<RegisterScreen> {
     return Scaffold(
       appBar: AppBar(
         title: const Center(
-          child: Text("Register"),
+          child: Text("Login"),
         ),
       ),
       body: Column(
         children: [
           Padding(
-            padding: const EdgeInsets.all(padding),
+            padding: const EdgeInsets.all(globals.padding),
             child: ElevatedButton(
-              onPressed: () => Navigator.pushReplacement(context,
-                  MaterialPageRoute(builder: (context) => const LoginScreen())),
-              child: const Text("Log in"),
+              onPressed: () => Navigator.pushReplacement(
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => const RegisterScreen())),
+              child: const Text("Register"),
             ),
           ),
           Padding(
-            padding: const EdgeInsets.all(padding),
-            child: TextField(
-              decoration: const InputDecoration(
-                border: OutlineInputBorder(),
-                labelText: "E-mail",
-              ),
-              onChanged: (value) {
-                setState(() {
-                  _email = value;
-                });
-              },
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(padding),
+            padding: const EdgeInsets.all(globals.padding),
             child: TextField(
               decoration: const InputDecoration(
                 border: OutlineInputBorder(),
@@ -62,7 +49,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
             ),
           ),
           Padding(
-            padding: const EdgeInsets.all(padding),
+            padding: const EdgeInsets.all(globals.padding),
             child: TextField(
               obscureText: true,
               decoration: const InputDecoration(
@@ -77,10 +64,14 @@ class _RegisterScreenState extends State<RegisterScreen> {
             ),
           ),
           Padding(
-            padding: const EdgeInsets.all(padding),
+            padding: const EdgeInsets.all(globals.padding),
             child: ElevatedButton(
-              onPressed: () {},
-              child: const Text("Register"),
+              onPressed: () => Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => ProfileScreen(
+                          email: "foo@bar.com", username: _username))),
+              child: const Text("Log in"),
             ),
           ),
         ],
