+++ lib/globals.dart
@@ -1,0 +1,1 @@
+const double padding = 10;
