+++ lib/main.dart
@@ -1,5 +1,7 @@
+import 'package:catppuccin_flutter/catppuccin_flutter.dart';
 import 'package:flutter/material.dart';
-import 'login.dart';
+import 'package:project_mobile_app/authentication/login.dart';
+import 'package:project_mobile_app/theme.dart';
 
 void main() {
   runApp(const App());
